Improve speed calculation, add bus selection and speed graph
- Rework speed smoothing: use speed averaging + endpoint cap instead of
  position centroid smoothing to prevent amplification bug
- Add stationary confirmation counter to avoid premature zero-speed
- Raise conservative factor from 0.92 to 0.95 (new smoothing is inherently conservative)
- Add bus selection in sidebar with speed graph panel
- Track buses in a map for proper stale-bus removal
- Show speed limits in bus list, sort by route then bus ID
- Set background color on index.html to avoid white flash
- Add verify_speed utility and regression tests

+++ shared/lib/src/services/geo_utils.dart
@@ -63,7 +63,7 @@ double pointToLineSegmentDistanceM(
 }
 
 /// Convert distance in meters and time delta in seconds to speed in km/h.
-double speedKmh(double distanceM, int timeDeltaS) {
+double speedKmh(double distanceM, double timeDeltaS) {
   if (timeDeltaS <= 0) return 0;
   return (distanceM / 1000) / (timeDeltaS / 3600);
 }
